Move Slack-specific icons back to their original folder
