+++ packages/react/src/modules/slack/components/SlackChannelCombobox/icons/CheckmarkIcon.tsx
