+++ packages/react/src/modules/slack/components/SlackChannelCombobox/icons/CloseIcon.tsx
