+++ packages/react/src/modules/slack/components/SlackChannelCombobox/icons/HashtagIcon.tsx
