+++ packages/react/src/modules/slack/components/SlackChannelCombobox/icons/InfoIcon.tsx
