+++ packages/react/src/modules/slack/components/SlackChannelCombobox/icons/LockIcon.tsx
