+++ packages/react/src/modules/slack/components/SlackChannelCombobox/icons/SearchIcon.tsx
