+++ packages/react/src/modules/slack/components/SlackChannelCombobox/icons/index.ts
